+++ Tekk/Views/ChatbotView.swift
@@ -3,47 +3,9 @@
 //  Tekk-frontend
 //
 //  Created by Jordan on 7/6/24.
-//
+//  This file contains the ChatbotView, which is used to display the chatbot.
 
 import SwiftUI
-
-// expected response structure from backend after GET request to conversations/<id> endpoint
-struct ConversationResponse: Codable {
-    let id: String
-    let messages: [APIMessage]
-}
-
-// expected response structure from backend after GET request to get_previous_conversations endpoint
-struct PreviousConversationsResponse: Codable {
-    let conversations: [Conversation]
-}
-
-// structure for messages in a conversation
-struct APIMessage: Codable {
-    let role: String
-    let content: String
-}
-
-// structure for chat messages
-struct Message_Struct: Identifiable {
-    let id = UUID()
-    let role: String
-    var content: String
-}
-
-// expected response structure from backend after POST request to generate_tutorial endpoint
-struct TutorialResponse: Codable {
-    let tutorial: String
-    let session_id: String
-    // let created_at: String
-}
-
-// structure for a conversation, displayed in ChatHistoryView to show previous conversations
-struct Conversation: Identifiable, Codable {
-    let id: String
-    var title: String
-    let createdAt: Date
-}
 
 // Main chatbot view of the app
 struct ChatbotView: View {
@@ -104,7 +66,7 @@ struct ChatbotView: View {
                     // Message input
                     HStack {
                         TextField("Let's get Tekky", text: $messageText)
-                            .padding()
+                            .padding() 
                             .background(Color.gray.opacity(0.1))
                             .cornerRadius(8)
                             .onSubmit {
@@ -154,6 +116,7 @@ struct ChatbotView: View {
         }
     }
     
+    // API call to send user input to the chatbot and get a response
     func sendMessage(message: String) {
         withAnimation {
             chatMessages.append(Message_Struct(role: "user", content: "[USER]" + message))
@@ -219,6 +182,7 @@ struct ChatbotView: View {
         .resume()
     }
 
+    // API call to load a conversation in some conversation session
     func loadConversation(_ id: String) {
         let url = URL(string: "http://127.0.0.1:8000/conversations/\(id)")!
         var request = URLRequest(url: url)
@@ -253,8 +217,8 @@ struct ChatbotView: View {
         .resume()
     }
 
+    // API call to start a new conversation (creates new conversation session ID and empty chat history)
     func startNewConversation() {
-        // API call to start a new conversation
         let url = URL(string: "http://127.0.0.1:8000/conversations/new")!
         var request = URLRequest(url: url)
         request.httpMethod = "POST"
